Compiling psetting.c on x86 with gcc generated a lot of warnings

+++ boot/psetting.c
@@ -201,7 +201,7 @@ static int cfReadINIFile(int argc, char *argv[])
 					int index=cfINIApps[cfINIApps_index].nkeys;
 					cfINIApps[cfINIApps_index].nkeys++;
 					memtmp = realloc(cfINIApps[cfINIApps_index].keys, sizeof(cfINIApps[cfINIApps_index].keys[0])*(index+1));
-					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #1 (%ld)\n", sizeof(cfINIApps[cfINIApps_index].keys[0])*(index+1)); _exit(1); }
+					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #1 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINIApps_index].keys[0])*(index+1))); _exit(1); }
 					cfINIApps[cfINIApps_index].keys = memtmp;
 					cfINIApps[cfINIApps_index].keys[index].key=NULL;
 					cfINIApps[cfINIApps_index].keys[index].str=NULL;
@@ -225,7 +225,7 @@ static int cfReadINIFile(int argc, char *argv[])
 					cfINIApps_index=cfINInApps;
 					cfINInApps++;
 					memtmp = realloc(cfINIApps, sizeof(cfINIApps[cfINInApps])*cfINInApps);
-					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #2 (%ld)\n", sizeof(cfINIApps[cfINInApps])*cfINInApps); _exit(1); }
+					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #2 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps])*cfINInApps)); _exit(1); }
 					cfINIApps = memtmp;
 					cfINIApps[cfINIApps_index].app=strdup(strbuf);
 					cfINIApps[cfINIApps_index].keys=NULL;
@@ -240,7 +240,7 @@ static int cfReadINIFile(int argc, char *argv[])
 					int index=cfINIApps[cfINIApps_index].nkeys;
 					cfINIApps[cfINIApps_index].nkeys++;
 					memtmp = realloc(cfINIApps[cfINIApps_index].keys, sizeof(cfINIApps[cfINIApps_index].keys[0])*(index+1));
-					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #3 (%ld)\n", sizeof(cfINIApps[cfINIApps_index].keys[0])*(index+1)); _exit(1); }
+					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #3 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINIApps_index].keys[0])*(index+1))); _exit(1); }
 					cfINIApps[cfINIApps_index].keys = memtmp;
 					cfINIApps[cfINIApps_index].keys[index].key=strdup(keybuf);
 					cfINIApps[cfINIApps_index].keys[index].str=strdup(strbuf);
@@ -264,7 +264,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 				cfINInApps++;
 				memtmp = realloc(cfINIApps, sizeof(cfINIApps[cfINInApps])*cfINInApps);
-				if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #4 (%ld)\n", sizeof(cfINIApps[cfINInApps])*cfINInApps); _exit(1); }
+				if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #4 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps])*cfINInApps)); _exit(1); }
 				cfINIApps = memtmp;
 				cfINIApps[cfINInApps-1].app=strdup("commandline__");
 				cfINIApps[cfINInApps-1].app[12]=argv[c][1];
@@ -284,7 +284,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 					cfINIApps[cfINInApps-1].nkeys++;
 					memtmp = realloc(cfINIApps[cfINInApps-1].keys, sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1));
-					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #5 (%ld)\n", sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1)); _exit(1); }
+					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #5 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1))); _exit(1); }
 					cfINIApps[cfINInApps-1].keys = memtmp;
 					cfINIApps[cfINInApps-1].keys[index].key=strdup("_");
 					cfINIApps[cfINInApps-1].keys[index].key[0]=*argvstat;
@@ -302,7 +302,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 		cfINInApps++;
 		memtmp = realloc(cfINIApps, sizeof(cfINIApps[cfINInApps])*cfINInApps);
-		if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #6 (%ld)\n", sizeof(cfINIApps[cfINInApps])*cfINInApps); _exit(1); }
+		if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #6 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps])*cfINInApps)); _exit(1); }
 		cfINIApps = memtmp;
 		cfINIApps[cfINInApps-1].app=strdup("CommandLine");
 		cfINIApps[cfINInApps-1].keys=NULL;
@@ -320,7 +320,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 				cfINIApps[cfINInApps-1].nkeys++;
 				memtmp = realloc(cfINIApps[cfINInApps-1].keys, sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1));
-				if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #7 (%ld)\n", sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1)); _exit(1); }
+				if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #7 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1))); _exit(1); }
 				cfINIApps[cfINInApps-1].keys = memtmp;
 				cfINIApps[cfINInApps-1].keys[index].key=strdup("_");
 				cfINIApps[cfINInApps-1].keys[index].key[0]=argv[c][1];
@@ -331,7 +331,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 		cfINInApps++;
 		memtmp = realloc(cfINIApps, sizeof(cfINIApps[cfINInApps])*cfINInApps);
-		if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #8 (%ld)\n", sizeof(cfINIApps[cfINInApps])*cfINInApps); _exit(1); }
+		if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #8 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps])*cfINInApps)); _exit(1); }
 		cfINIApps = memtmp;
 		cfINIApps[cfINInApps-1].app=strdup("CommandLine--");
 		cfINIApps[cfINInApps-1].keys=NULL;
@@ -349,7 +349,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 				cfINIApps[cfINInApps-1].nkeys++;
 				memtmp = realloc(cfINIApps[cfINInApps-1].keys, sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1));
-				if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #9 (%ld)\n", sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1)); _exit(1); }
+				if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #9 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1))); _exit(1); }
 				cfINIApps[cfINInApps-1].keys = memtmp;
 				cfINIApps[cfINInApps-1].keys[index].key=strdup(argv[c]+2);
 				cfINIApps[cfINInApps-1].keys[index].str=strdup("1");
@@ -359,7 +359,7 @@ static int cfReadINIFile(int argc, char *argv[])
 
 		cfINInApps++;
 		memtmp = realloc(cfINIApps, sizeof(cfINIApps[cfINInApps])*cfINInApps);
-		if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #10 (%ld)\n", sizeof(cfINIApps[cfINInApps])*cfINInApps); _exit(1); }
+		if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #10 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps])*cfINInApps)); _exit(1); }
 		cfINIApps = memtmp;
 		cfINIApps[cfINInApps-1].app=strdup("CommandLine_Files");
 		cfINIApps[cfINInApps-1].keys=NULL;
@@ -396,7 +396,7 @@ static int cfReadINIFile(int argc, char *argv[])
 						sprintf(buffer, "playlist%d", playlists++);
 					cfINIApps[cfINInApps-1].nkeys++;
 					memtmp = realloc(cfINIApps[cfINInApps-1].keys, sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1));
-					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #11 (%ld)\n", sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1)); _exit(1); }
+					if (!memtmp) { fprintf (stderr, "cfReadINIFile() realloc failed #11 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps-1].keys[0])*(index+1))); _exit(1); }
 					cfINIApps[cfINInApps-1].keys = memtmp;
 					cfINIApps[cfINInApps-1].keys[index].key=strdup(buffer);
 					cfINIApps[cfINInApps-1].keys[index].str=strdup(argv[c]+(argv[c][0]=='@'?1:0));
@@ -475,7 +475,7 @@ doappend:
 			j=cfINIApps[i].nkeys;
 			cfINIApps[i].nkeys++;
 			memtmp = realloc(cfINIApps[i].keys, sizeof(cfINIApps[i].keys[0])*(j+1));
-			if (!memtmp) { fprintf (stderr, "cfSetProfileString() realloc failed #1 (%ld)\n", sizeof(cfINIApps[i].keys[0])*(j+1)); _exit(1); }
+			if (!memtmp) { fprintf (stderr, "cfSetProfileString() realloc failed #1 (%lu)\n", (unsigned long)(sizeof(cfINIApps[i].keys[0])*(j+1))); _exit(1); }
 			cfINIApps[i].keys = memtmp;
 			cfINIApps[i].keys[j].key=strdup(key);
 			cfINIApps[i].keys[j].str=strdup(str);
@@ -485,7 +485,7 @@ doappend:
 		}
 	cfINInApps++;
 	memtmp = realloc(cfINIApps, sizeof(cfINIApps[cfINInApps])*cfINInApps);
-	if (!memtmp) { fprintf (stderr, "cfSetProfileString() realloc failed #2 (%ld)\n", sizeof(cfINIApps[cfINInApps])*cfINInApps); _exit(1); }
+	if (!memtmp) { fprintf (stderr, "cfSetProfileString() realloc failed #2 (%lu)\n", (unsigned long)(sizeof(cfINIApps[cfINInApps])*cfINInApps)); _exit(1); }
 	cfINIApps = memtmp;
 	cfINIApps[i].app=strdup(app);
 	cfINIApps[i].keys=NULL;
